+++ F2812_eZdsp_XDS100usb/Source/Dec_Ini_Vars.c
@@ -1,5 +1,6 @@
 //Variables initialization
 #include "DSP281x_Device.h"
 
-int dummy1 = 0;
+int16 dummy1 = 0;
+
